# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Mother Bot API - SKZ Financial Hub
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List withdrawals
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalListResponse"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Create a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWithdrawalBody"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Invalid request or insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [withdrawals]
      summary: Approve a withdrawal request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveWithdrawalBody"
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /withdrawals/{id}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [withdrawals]
      summary: Reject a withdrawal request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectWithdrawalBody"
      responses:
        "200":
          description: Withdrawal rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BotApiKey:
      type: apiKey
      in: header
      name: X-Bot-Api-Key
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        currency:
          type: string
        amount:
          type: string
        fee:
          type: string
        netAmount:
          type: string
        method:
          type: string
        address:
          type: ["string", "null"]
        txHash:
          type: ["string", "null"]
        status:
          type: string
        rejectedReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - currency
        - amount
        - fee
        - netAmount
        - method
        - status
        - createdAt
    CreateWithdrawalBody:
      type: object
      properties:
        userId:
          type: integer
        currency:
          type: string
        amount:
          type: string
        method:
          type: string
        address:
          type: string
      required:
        - userId
        - currency
        - amount
        - method
    ApproveWithdrawalBody:
      type: object
      properties:
        txHash:
          type: string
      required:
        - txHash
    RejectWithdrawalBody:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
    WithdrawalListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...